refactor: useSession 타입 분리
